Fixed several ui label mistakes and fixed conversion errors
Several ui labels for the ports were mislabeled and have now been corrected.
Added a tag to the end of each of the channel labels denoting the port they are connected to (I.e. P1 for port 1 etc)
Added a new conversion (Conversion 0) to represent simply keeping the raw voltages and updated where the data is converted and the desired conversion selected to incorperate this now.